# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FMS Log Nexus API
  description: |
    Centralized logging and monitoring platform API for FMS Log Nexus.
    
    ## Authentication
    
    The API supports two authentication methods:
    - **JWT Bearer Token**: For dashboard users and interactive sessions
    - **API Key**: For service accounts and job integrations
    
    ## Rate Limiting
    
    - Log ingestion: 10,000 requests/minute per API key
    - Dashboard queries: 1,000 requests/minute per user
    
  version: 1.0.0
  contact:
    name: FMS Log Nexus Support
    email: support@company.com
  license:
    name: Proprietary
servers:
  - url: https://lognexus.company.com/api
    description: Production server
  - url: https://localhost:5001/api
    description: Development server
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alert definitions
      responses:
        '200':
          description: Alert definitions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'

    post:
      tags: [Alerts]
      summary: Create alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertCreateRequest'
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/{id}:
    get:
      tags: [Alerts]
      summary: Get alert details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetail'

    put:
      tags: [Alerts]
      summary: Update alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertUpdateRequest'
      responses:
        '200':
          description: Alert updated

    delete:
      tags: [Alerts]
      summary: Delete alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Alert deleted
  /alerts/instances:
    get:
      tags: [Alerts]
      summary: List alert instances
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [New, Acknowledged, Resolved]
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/AlertSeverity'
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Alert instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertInstancesResponse'
  /alerts/instances/{id}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeRequest'
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertInstanceStatusResponse'
  /alerts/instances/{id}/resolve:
    post:
      tags: [Alerts]
      summary: Resolve alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveRequest'
      responses:
        '200':
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertInstanceStatusResponse'

  # ============================================================================
  # System Endpoints
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token for user authentication

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
      description: Items per page
...
    StartDate:
      name: startDate
      in: query
      schema:
        type: string
        format: date-time
      description: Start of time range (ISO 8601)
    EndDate:
      name: endDate
      in: query
      schema:
        type: string
        format: date-time
      description: End of time range (ISO 8601)
  schemas:
...
    AlertSeverity:
      type: string
      enum: [Low, Medium, High, Critical]
    AlertType:
      type: string
      enum: [ErrorThreshold, JobFailure, ServerOffline, PerformanceWarning, CustomQuery, PatternMatch]

    # ========== Common ==========
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        totalCount:
          type: integer
        hasNextPage:
          type: boolean
        hasPreviousPage:
          type: boolean
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        alertType:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        isActive:
          type: boolean
        throttleMinutes:
          type: integer
        lastTriggeredAt:
          type: string
          format: date-time
        triggerCount:
          type: integer
    AlertDetail:
      allOf:
        - $ref: '#/components/schemas/Alert'
        - type: object
          properties:
            condition:
              type: object
            notificationChannels:
              type: object
            jobId:
              type: string
            serverName:
              type: string
            createdAt:
              type: string
              format: date-time
            createdBy:
              type: string
    AlertCreateRequest:
      type: object
      required: [name, alertType, severity, condition]
      properties:
        name:
          type: string
        description:
          type: string
        alertType:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        condition:
          type: object
        throttleMinutes:
          type: integer
          default: 15
        notificationChannels:
          type: object
        jobId:
          type: string
        serverName:
          type: string
        isActive:
          type: boolean
          default: true
    AlertUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        condition:
          type: object
        throttleMinutes:
          type: integer
        notificationChannels:
          type: object
        isActive:
          type: boolean
    AlertsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
    AlertInstance:
      type: object
      properties:
        id:
          type: string
          format: uuid
        alertId:
          type: string
          format: uuid
        alertName:
          type: string
        triggeredAt:
          type: string
          format: date-time
        message:
          type: string
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        status:
          type: string
          enum: [New, Acknowledged, Resolved, Suppressed]
        acknowledgedAt:
          type: string
          format: date-time
        acknowledgedBy:
          type: string
        resolvedAt:
          type: string
          format: date-time
        resolvedBy:
          type: string
    AlertInstancesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AlertInstance'
        summary:
          type: object
          properties:
            new:
              type: integer
            acknowledged:
              type: integer
            resolved:
              type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    AcknowledgeRequest:
      type: object
      properties:
        note:
          type: string
    ResolveRequest:
      type: object
      properties:
        note:
          type: string
    AlertInstanceStatusResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        acknowledgedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time

    # ========== System ==========
...